feat: migrate fonts to next/font, extract to lib/fonts.ts
- Replace geist package + next/font/local with next/font/google for Geist
  (weights 400/500/600/700, variable font)
- Refresh CalSans-SemiBold.woff2 from Google Fonts CDN; drop .ttf and .woff
- Keep Cal Sans on next/font/local declared at weight 600 to match prior setup
- Extract all font config to lib/fonts.ts, export fontVariables (matching buddy pattern)
- Remove geist package dependency
